encore un petit pas, de petite modif

+++ src/main/java/com/istat/freedev/processor/interfaces/RunnableDispatcher.java
@@ -1,7 +1,7 @@
 package com.istat.freedev.processor.interfaces;
 
 public interface RunnableDispatcher {
-    void dispatch(Runnable runnable, int delay);
+    void dispatch(Runnable runnable, long delay);
 
     void cancel(Runnable runnable);
